Disable RedHat dnf plugins when migrating a RHEL installation
Because the migration script does not remove subscription manager,
RedHat-specific dnf plugins are left installed when migrating a
RHEL system to AlmaLinux, even though the plugins cannot be used
on AlmaLinux.

This commit modifies the migration script so that it disables
RedHat-specific dnf plugins before attempting to migrate the system.